refactored:reafctored the code according to front end

+++ app/models/note.model.js
@@ -3,22 +3,25 @@ const mongoose = require('mongoose');
 const NoteSchema = mongoose.Schema({
     title: String,
     content: String,
-    postedBy: {
-        type: mongoose.Schema.Types.ObjectId,
-        userId: { type: mongoose.Schema.Types.ObjectId, ref: "User" },
-    }
+     userId: { 
+         type: mongoose.Schema.Types.ObjectId,
+          ref: "User" },
+   
 }, { timestamps: true });
 const Note = mongoose.model('Note', NoteSchema);
 // Create a Note
-const createNote = (title, content) => {
-    const note = new Note({ title: title, content: content });
+const createNote = (title, content,userId) => {
+    const note = new Note({ title: title, content: content,userId:userId });
     // Save Note in the database
     return note.save()
 
 };
 // find all notes
-const findAllNotes = () => {
-    return Note.find().populate('userId');
+const findAllNotes = (userId) => {
+    return Note.find({userId:userId}).populate({
+        path:"userId",
+        select:["firstname","lastname","email"]
+    });
 }
 // query to find a single note
 const findNote = (userId,findId, callback) => {
@@ -27,8 +30,8 @@ const findNote = (userId,findId, callback) => {
     })
 }
 // Find note and update it with the request body
-const updateNote = (userid,findId, title, content) => {
-    return Note.findOneAndUpdate({userId:userid,findId:findId}, {
+const updateNote = (userId,findId, title, content) => {
+    return Note.findOneAndUpdate({userId:userId,findId:findId}, {
         title: title,
         content: content
     }, { new: true }).then((result) => {
@@ -40,8 +43,8 @@ const updateNote = (userid,findId, title, content) => {
     })
 }
 // query to delete a note
-const deleteById = (findId) => {
-    return Note.findByIdAndRemove(findId)
+const deleteById = (userId,findId) => {
+    return Note.findByIdAndRemove({userId:userId,findId:findId})
 }
 module.exports = {
     createNote,
